ref: move bitbucket testutils to fixtures
