+++ fixtures/bitbucket.py
